keep-alive race condition and rate limit retry
Fixed 3 related but separate retry scenarios
1. the rate-limit/backoff retry scenario did not work if the request contained a body because the body stream had not been reset prior to the retry (PUT/POST)
There is a keep-alive race condition that exists when the server disconnects idle connections. If a request is made in the window in between when the server has disconnected, but before the client has detected the disconnect, it looks to the client as if the server returned a blank HTTP status line to the sent request. Idempotent methods (PUT, DELETE, GET) are retried automatically, however, POST is not.
2. A PUT request retry attempt suffers from the same problem as #1 above (body content stream not reset). That will raise a NonRepeatableRequestException that is now handled (by resetting the body content stream).
3. A POST request will not be retried automatically by the client, so handle NoHttpResponseException exceptions and retry POST after resetting the body content stream.